[CMS][m]: Refactor WP CMS backend as plugin
