+++ env.template
@@ -1,4 +1,4 @@
-PLUGINS="example"
+PLUGINS="example wp"
 WP_TOKEN=
 NODE_ENV=development
 THEME=example
