+++ tests/lib/index.test.js
@@ -1,13 +1,13 @@
 const test = require('ava')
 
-const cms = require('../../lib/cms')
+const wp = require('../../plugins/wp/cms')
 const dms = require('../../lib/dms')
 const config = require('../../config')
 const mocks = require('../../fixtures')
 
 mocks.initMocks()
 
-const CmsModel = new cms.CmsModel()
+const WPModel = new wp.CmsModel()
 const DmsModel = new dms.DmsModel(config)
 
 
@@ -15,7 +15,7 @@ test('getPost api works', async t => {
   t.plan(1)
 
   const slug = 'about'
-  const result = await CmsModel.getPost(slug)
+  const result = await WPModel.getPost(slug)
 
   t.is(result.title, 'Welcome to Data Service!')
 })
@@ -24,7 +24,7 @@ test('getPost api works', async t => {
 test('getListOfPosts api works', async t => {
   t.plan(1)
 
-  const result = await CmsModel.getListOfPosts()
+  const result = await WPModel.getListOfPosts()
 
   t.is(result.length, 2)
 })
